requirements.txt
PySide6
imagehash>=4.3.0
Pillow>=9.0.0
send2trash>=1.8.0
psutil>=5.9.0
